+++ examples/preset.rs
@@ -1,19 +1,11 @@
-use modtype::ConstValue;
+use modtype::use_modtype;
 
-#[derive(Clone, Copy, PartialEq, Eq, PartialOrd, Ord, Debug, ConstValue)]
-#[modtype(const_value = 1_000_000_007u64)]
-enum M {}
-
-type Z = modtype::u64::Z<M>;
-
-#[allow(non_snake_case)]
-fn Z(value: u64) -> Z {
-    Z::new(value)
-}
+#[use_modtype]
+type F = modtype::u64::F<1_000_000_007u64>;
 
 fn main() {
     static INPUT: &str = "13";
-    let mut a = INPUT.parse::<Z>().unwrap();
-    a += Z(1_000_000_000);
+    let mut a = INPUT.parse::<F>().unwrap();
+    a += F(1_000_000_000);
     dbg!(a);
 }