+++ tests/pow.rs
@@ -1,4 +1,4 @@
-use modtype::ConstValue;
+use modtype::use_modtype;
 
 use num::pow::Pow as _;
 use num::BigUint;
@@ -6,16 +6,8 @@ use once_cell::sync::Lazy;
 
 #[test]
 fn it_works() {
-    #[derive(Clone, Copy, PartialEq, Eq, PartialOrd, Ord, Debug, ConstValue)]
-    #[modtype(const_value = 1_000_000_007u64)]
-    enum M {}
-
-    type F = modtype::u64::F<M>;
-
-    #[allow(non_snake_case)]
-    fn F(value: u64) -> F {
-        F::new(value)
-    }
+    #[use_modtype]
+    type F = modtype::u64::F<1_000_000_007u64>;
 
     const BASE: u64 = 123_456_789;
     const EXP: u32 = 123;